refactor(analysis-types): move churn serde tests
